write code to handle form submit
