+++ web/package.json
@@ -18,11 +18,13 @@
     "react-dom": "^17.0.1",
     "react-icons": "^4.1.0",
     "react-input-mask": "^2.0.4",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "react-select": "^3.2.0",
     "reactjs-popup": "^2.0.4",
     "styled-components": "^5.2.1",
     "typescript": "^4.0.3",
+    "uuid": "^8.3.2",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
@@ -52,7 +54,9 @@
   },
   "devDependencies": {
     "@types/react-input-mask": "^3.0.0",
+    "@types/react-router-dom": "^5.1.7",
     "@types/react-select": "^3.1.2",
-    "@types/styled-components": "^5.1.7"
+    "@types/styled-components": "^5.1.7",
+    "@types/uuid": "^8.3.0"
   }
 }
