+++ web/src/components/PlaceCard/styles.ts
@@ -82,3 +82,9 @@ export const CardContent = styled.div`
     }
   }
 `;
+
+export const ModalFormContainer = styled.div``;
+export const Form = styled.div``;
+export const Select = styled.div``;
+export const Input = styled.div``;
+
